% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{baks}
\alias{baks}
\title{Exponential smoothing function.}
\usage{
baks(spikes, t, a = 4, b = 0.975)
}
\arguments{
\item{spikes}{A binarized vector (0 and 1) of spikes}

\item{t}{A vector of times in seconds, same length as spikes}

\item{a}{shape parmeter according to Ahmadi et al., defaults to 4}

\item{b}{scale parameter according to Ahmadi et al., defaults to 0.975}

\item{y}{Your y values (measures ?)}

\item{start}{: The first x value}

\item{stop}{: The last x value}

\item{na.stop}{: returns NA if one value is NA}

\item{ex.na.stop}{: returns NA if the first or the last value is NA
Function to perform exponential smoothing.}

\item{x}{}

\item{alpha}{}
}
\value{


A vector of firing rate density estimation
}
\description{
Calculate the Area Under Curve of y~x
}
\references{
https://www.itl.nist.gov/div898/software/dataplot/refman2/auxillar/exposmoo.htm
Helper function to calculate a firing rate estimate using a Bayesian Adaptive Kernel Smoother as suggested by Ahmadi et al. 2018
https://doi.org/10.1371/journal.pone.0206794
}
\author{
Stephen Eglen

Muad Abd El Hay
}
