% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculators.R
\name{binarize}
\alias{binarize}
\title{Estimate spikes and binarize data using the L0 method by Jewell et al. 2019}
\usage{
binarize(cexp, slot = "normalized", ...)
}
\arguments{
\item{cexp}{CalciumExperiment object to binarize.}

\item{slot}{The assay slot to use for binarization. Defaults to "normalized".}

\item{...}{Additional parameters that can be passed to the spike estimation function.}
}
\value{
A CalciumExperiment object with an additional assay containig the estimated spikes as 0 and 1 binary.
}
\description{
Function to estimate spikes and binarize a CalciumExperiment object.
}
\author{
Muad Abd El Hay
}
