% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CalciumExperiment-class}
\alias{CalciumExperiment-class}
\alias{.CalciumExperiment}
\title{}
\arguments{
\item{raw}{}

\item{...}{}
}
\value{

}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Muad Abd El Hay
#' @export
}
