% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{CalciumExperiment}
\alias{CalciumExperiment}
\title{}
\usage{
CalciumExperiment(raw, ...)
}
\arguments{
\item{raw}{}

\item{...}{}
}
\value{

}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Muad Abd El Hay
}
