% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{colScale}
\alias{colScale}
\title{Fast center and/or scale columns using the matrixStats functions.}
\usage{
colScale(
  x,
  center = TRUE,
  scale = TRUE,
  add_attr = TRUE,
  rows = NULL,
  cols = NULL
)
}
\arguments{
\item{x}{}

\item{center}{}

\item{scale}{}

\item{add_attr}{}

\item{rows}{}

\item{cols}{}
}
\value{
A matrix with scaled columns.
}
\description{
Function for fast center and scaling of matrix columns.
}
\references{
https://hopstat.wordpress.com/2016/02/23/a-faster-scale-function/
}
