% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{estimateBaselines}
\alias{estimateBaselines}
\title{Function to estimate baselines of traces.}
\usage{
estimateBaselines(x, method = "irls", ...)
}
\arguments{
\item{x}{}

\item{method}{}

\item{...}{}
}
\value{

}
\description{
Function to estimate baselines using the baseline package. Default method is "irls". For other options refer to the baseline function in the baseline package.
}
\author{
Muad Abd El Hay
}
