% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculators.R
\name{estimateFiringRate}
\alias{estimateFiringRate}
\title{Firing rate estimation from binarized data}
\usage{
estimateFiringRate(cexp, slot = "l0spikes", ...)
}
\arguments{
\item{cexp}{CalciumExperiment object that underwent binarization}

\item{slot}{Assay slot containing binarized data}

\item{...}{Variables passed to the bayesian adaptive kernel smoother (see \link{baks} for more information).}
}
\value{
A CalciumExperiment object with firing rate estimation
}
\description{
A wrapper function for firing rate estimation of binarized CalciumExperiment objects.
}
\author{
Muad Abd El Hay
}
