% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{expSmooth}
\alias{expSmooth}
\title{}
\usage{
expSmooth(x, alpha = 0.3)
}
\arguments{
\item{x}{}

\item{alpha}{}
}
\value{

}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\references{
https://www.itl.nist.gov/div898/software/dataplot/refman2/auxillar/exposmoo.htm
}
\author{
Stephen Eglen
}
