% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extractSpikesFromLZeroFit}
\alias{extractSpikesFromLZeroFit}
\title{Fit L0 spike inference model and extract spikes according to Jewell et al. 2019}
\usage{
extractSpikesFromLZeroFit(x, gam = 0.8, lambda = 0.1, constraint = FALSE)
}
\arguments{
\item{x}{A vector containing the normalized fluorescence trace.}

\item{gam}{numeric. between 0 and 1. Parameter to tune spike inference.}

\item{lambda}{numeric. between 0 and 1. Parameter to tune spike inference.}

\item{constraint}{logical. Whether to run the fit with constrain or not.}
}
\value{
A vector containig the estimated spikes as 1s and 0s.
}
\description{
Utility function to extract the estimated spikes based on the FastLZeroSpikeInference package by Jewell et al. 2019.
}
\author{
Muad Abd El Hay
}
