% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculators.R
\name{getStimulusParameters}
\alias{getStimulusParameters}
\title{Calculate base parameters.}
\usage{
getStimulusParameters(
  cexp,
  slot = "normalized",
  binslot = "l0spikes",
  stimulus.trace,
  ...
)
}
\arguments{
\item{cexp}{CalciumExperiment object that has base parameters calculated already. Should be normalized and binarized before running.}

\item{slot}{The assay slot representing the normalized calcium traces}

\item{binslot}{The assay slot with the binarized traces}

\item{stimulus.trace}{The rowData column containing the stimulus to calculate for}

\item{...}{}
}
\value{
A CalciumExperiment object with calculated base parameters added to the colData DataFrame
}
\description{
Calculates stimulus-dependent parameters of a CalciumExperiment object such as stimulus threshold, correlation, and mutual information.
}
\author{
Muad Abd El Hay
}
