% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{normalize,CalciumExperiment-method}
\alias{normalize,CalciumExperiment-method}
\title{Function to normalize traces.}
\usage{
\S4method{normalize}{CalciumExperiment}(
  object,
  method = "estimateBaseline",
  slot,
  window,
  output.name = "normalized",
  ...
)
}
\arguments{
\item{object}{CalciumExperiment object.}

\item{method}{character. The method to use for baseline correction. Defaults to "estimateBaseline",}

\item{slot}{character. Which assay slot to use for baseline correction.}

\item{window}{numeric. The width of the baseline window to use for dF/F0 calculation. A window of width 500 will take the first 500 frames as baseline.}

\item{...}{}
}
\value{
A CalciumExperiment object with a "normalized" assay.
}
\description{
Function to normalize traces.
}
