% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{plotSpikes}
\alias{plotSpikes}
\title{Function to plot spikes from CalciumExperiment objects.}
\usage{
plotSpikes(x, slot = "l0spikes", order = FALSE, cluster = FALSE, scale = TRUE)
}
\arguments{
\item{x}{}

\item{slot}{chracter. The assays() slot to use for plotting.}

\item{order}{logical. Whether to order the cells/ROIs by calculating the sums of the response.}

\item{cluster}{logical. Whether to clusted the cells/ROIs by correlation.}
}
\value{
A ggplot2 object with the predicted pikes for each cell.
}
\description{
Plots staggered barcode plots of each ROI. Defaults to "l0spikes" values without any additional modifications.
}
\details{
Can be set to order the traces or even cluster them before plotting. Traces are scaled by default.
}
\author{
Muad Abd El Hay
}
