% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{plotTraces}
\alias{plotTraces}
\title{}
\usage{
plotTraces(x, slot = "raw", order = FALSE, cluster = FALSE)
}
\arguments{
\item{x}{}

\item{slot}{chracter. The assays() slot to use for plotting.}

\item{order}{logical. Whether to order the cells/ROIs by calculating the sums of the response.}

\item{cluster}{logical. Whether to clusted the cells/ROIs by correlation.}
}
\value{

}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Muad Abd El Hay
}
