% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{plotTraces2}
\alias{plotTraces2}
\title{Function to plot CalciumExperiment objects.}
\usage{
plotTraces2(x, slot = "raw", order = FALSE, cluster = FALSE, scale = TRUE)
}
\arguments{
\item{x}{}

\item{slot}{chracter. The assays() slot to use for plotting.}

\item{order}{logical. Whether to order the cells/ROIs by calculating the sums of the response.}

\item{cluster}{logical. Whether to clusted the cells/ROIs by correlation.}
}
\value{
A ggplot2 object with the traces for each cell.
}
\description{
Plots staggered line plots of each ROI. Defaults to "raw" values without any additional modifications.
}
\details{
Can be set to order the traces or even cluster them before plotting. Traces are scaled by default.
}
\author{
Muad Abd El Hay
}
