% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{randomAlphaNum}
\alias{randomAlphaNum}
\title{Create a random alpha-numeric string of length n.}
\usage{
randomAlphaNum(n = 1)
}
\arguments{
\item{n}{Length of the alphanumeric string.}
}
\value{
A string with random alpha-numerics of length n.
}
\description{
Helper function to create an assay name when none is supplied.
}
\author{
Muad Abd El Hay
}
