% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{raw,CalciumExperiment-method}
\alias{raw,CalciumExperiment-method}
\title{}
\usage{
\S4method{raw}{CalciumExperiment}(x, withDimnames = TRUE)
}
\arguments{
\item{x}{}

\item{withDimnames}{}
}
\value{

}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Muad Abd El Hay
}
