% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{raw}
\alias{raw}
\title{}
\usage{
raw(x, ...)
}
\arguments{
\item{x}{}

\item{...}{}
}
\value{

}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Muad Abd El Hay
}
