% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importers.R
\name{readSuite2p}
\alias{readSuite2p}
\title{Import suite2p outputs}
\usage{
readSuite2p(s2p.path, time, exp.n, correction.factor = 0.7, ...)
}
\arguments{
\item{s2p.path}{The path to the folder containing suite2p outputs}

\item{time}{A time vector giving the timepoint at which every frame was taken. If left blank, a vector will be calculated with the fs parameter from Suite2p and the number of frames.}

\item{exp.n}{Experiment name/number. This is used to give the cells/ROIs unique IDs and will be randomly generated if left blank.}

\item{correction.factor}{The corection factor for neuropili substraction, defaults to 0.7}

\item{...}{}
}
\value{
A CalciumExperiment object with the data from a Suite2p output folder.
}
\description{
Load data from suite2p
}
\details{
Takes the output of a full suite2p run and turns it into a \linkS4class{CalciumExperiment} object
}
\author{
Muad Abd El Hay
}
