% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{splitByStimulus}
\alias{splitByStimulus}
\title{Split CalciumExperiment object into multiple assays by stimuli}
\usage{
splitByStimulus(
  cexp,
  stim.table,
  stim.table.type = "time",
  buffer = FALSE,
  buffer.size = 0.5,
  ...
)
}
\arguments{
\item{cexp}{CalciumExperiment to split.}

\item{stim.table}{The stimulus table with columns "stimulus", "begin", "end" where every row is one stimulus to split the dataset by.}

\item{stim.table.type}{The type of stimulus table that is supplied.}

\item{buffer}{Whether to add a "buffer" before and after the stimulus. Useful for plotting.}

\item{buffer.size}{The relative size of the buffer. Defaults to half of the stimulus length.}

\item{...}{}
}
\value{
A MultiAssayExperiment object with each stimulus as independent CalciumExperiment object.
}
\description{
Takes as input a CalciumExperiment object and a stimulus table that must contain the columns "stimulus", "begin", and "end" denoting the name/number of the stimulus, the beginning of the stimulus and the end, respectively. The stimulus table could either denote the times of stimuli using the units sipplied in the "time" field (check rowData of the CalciumExperiment object) or contain the frame numbers (row numbers) of the stimulus start and end. The default assumes a table contining the times.
}
\author{
Muad Abd El Hay
}
