% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CalciumExperiment-class}
\alias{CalciumExperiment-class}
\alias{.CalciumExperiment}
\title{CalciumExperiment class}
\arguments{
\item{raw}{}

\item{...}{}
}
\value{

}
\description{
Function that defines the CalciumExperiment class
}
\author{
Muad Abd El Hay
}
