% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{CalciumExperiment}
\alias{CalciumExperiment}
\title{CalciumExperiment constructor}
\usage{
CalciumExperiment(raw, ...)
}
\arguments{
\item{raw}{}

\item{...}{}
}
\value{
A CalciumExperiment object.
}
\description{
Constructor for CalciumExperiment objects.
}
\author{
Muad Abd El Hay
}
