% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{expSmooth}
\alias{expSmooth}
\title{Exponential smoothing function.}
\usage{
expSmooth(x, alpha = 0.3)
}
\arguments{
\item{x}{}

\item{alpha}{}
}
\value{

}
\description{
Function to perform exponential smoothing.
}
\references{
https://www.itl.nist.gov/div898/software/dataplot/refman2/auxillar/exposmoo.htm
}
\author{
Stephen Eglen
}
