% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{raw,CalciumExperiment-method}
\alias{raw,CalciumExperiment-method}
\title{raw Accessor}
\usage{
\S4method{raw}{CalciumExperiment}(x, withDimnames = TRUE)
}
\arguments{
\item{x}{}

\item{withDimnames}{}
}
\value{
The raw fluorescence trace of a CalciumExperiment object.
}
\description{
Accessor function for raw fluorescence values per ROI
}
\author{
Muad Abd El Hay
}
