% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{raw}
\alias{raw}
\title{raw Accessor}
\usage{
raw(x, ...)
}
\arguments{
\item{x}{}

\item{...}{}
}
\value{

}
\description{
Accessor function for raw fluorescence values per ROI
}
\author{
Muad Abd El Hay
}
